Function Index

List of functions in TrialDocs module

TrialDocs.domath — Method.

domath(x::Number)

Return x + 5.

source

TrialDocs.hello — Method.

hello(who::String)

Return "Hello, who".

source

TrialDocs.test_args — Method.

test_args(args...)

Description

Testing to see if I can create documentation from a different folder

Arguments

args...: (arbitrary as long as it works with println).

Return

Nothing

Additional details

source

TrialDocs.test_args_kw — Method.

testargskw(args..., keyword = "keyword")

Description

Testing to see if I can create documentation from a different folder

Arguments

args...: (arbitrary as long as it works with println).

Keyword Arguments

keyword: default = "keyword"

Return

Nothing

Additional details

Just another test

source